fix echo and generic aea tests

+++ tests/test_packages/test_skills/test_echo.py
@@ -58,9 +58,10 @@ class TestEchoSkill:
         """Run the echo skill sequence."""
         if pytestconfig.getoption("ci"):
             pytest.skip("Skipping the test since it doesn't work in CI.")
+
         # add packages folder
         packages_src = os.path.join(self.cwd, "packages")
-        packages_dst = os.path.join(os.getcwd(), "packages")
+        packages_dst = os.path.join(self.t, "packages")
         shutil.copytree(packages_src, packages_dst)
 
         # create agent
@@ -95,9 +96,9 @@ class TestEchoSkill:
             stdout=subprocess.PIPE,
             env=os.environ.copy(),
         )
+        time.sleep(2.0)
 
         # add sending and receiving envelope from input/output files
-        time.sleep(2.0)
         message = DefaultMessage(type=DefaultMessage.Type.BYTES, content=b"hello")
         expected_envelope = Envelope(
             to=self.agent_name,
@@ -112,44 +113,44 @@ class TestEchoSkill:
             expected_envelope.message.decode("utf-8"),
         )
         encoded_envelope = encoded_envelope.encode("utf-8")
-        with open(Path(self.t, self.agent_name, "input_file"), "ab+") as f:
-            f.write(encoded_envelope + b"\n")
-            f.flush()
+        try:
+            with open(Path(self.t, self.agent_name, "input_file"), "ab+") as f:
+                f.write(encoded_envelope + b"\n")
+                f.flush()
 
-        time.sleep(2.0)
-        with open(Path(self.t, self.agent_name, "output_file"), "rb+") as f:
-            lines = f.readlines()
+            time.sleep(2.0)
+            with open(Path(self.t, self.agent_name, "output_file"), "rb+") as f:
+                lines = f.readlines()
 
-        assert len(lines) == 1
-        line = lines[0]
-        to, sender, protocol_id, message = line.strip().split(b",", maxsplit=3)
-        to = to.decode("utf-8")
-        sender = sender.decode("utf-8")
-        protocol_id = PublicId.from_str(protocol_id.decode("utf-8"))
+            assert len(lines) == 1
+            line = lines[0]
+            to, sender, protocol_id, message = line.strip().split(b",", maxsplit=3)
+            to = to.decode("utf-8")
+            sender = sender.decode("utf-8")
+            protocol_id = PublicId.from_str(protocol_id.decode("utf-8"))
 
-        actual_envelope = Envelope(
-            to=to, sender=sender, protocol_id=protocol_id, message=message
-        )
-        assert expected_envelope.to == actual_envelope.sender
-        assert expected_envelope.sender == actual_envelope.to
-        assert expected_envelope.protocol_id == actual_envelope.protocol_id
-        assert expected_envelope.message == actual_envelope.message
+            actual_envelope = Envelope(
+                to=to, sender=sender, protocol_id=protocol_id, message=message
+            )
+            assert expected_envelope.to == actual_envelope.sender
+            assert expected_envelope.sender == actual_envelope.to
+            assert expected_envelope.protocol_id == actual_envelope.protocol_id
+            assert expected_envelope.message == actual_envelope.message
+            time.sleep(2.0)
+        finally:
+            process.send_signal(signal.SIGINT)
+            process.wait(timeout=20)
+            if not process.returncode == 0:
+                poll = process.poll()
+                if poll is None:
+                    process.terminate()
+                    process.wait(2)
 
-        time.sleep(2.0)
-        process.send_signal(signal.SIGINT)
-        process.wait(timeout=20)
-
-        assert process.returncode == 0
-
-        poll = process.poll()
-        if poll is None:
-            process.terminate()
-            process.wait(2)
-
-        os.chdir(self.t)
-        self.result = self.runner.invoke(
-            cli, [*CLI_LOG_OPTION, "delete", self.agent_name], standalone_mode=False
-        )
+            os.chdir(self.t)
+            result = self.runner.invoke(
+                cli, [*CLI_LOG_OPTION, "delete", self.agent_name], standalone_mode=False
+            )
+            assert result.exit_code == 0
 
     @classmethod
     def teardown_class(cls):
